+++ project-keeper/src/test/java/com/exasol/projectkeeper/ProjectKeeperAbstractIT.java
@@ -5,46 +5,28 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-import java.io.IOException;
 import java.nio.file.Path;
+import java.util.Optional;
 
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.io.TempDir;
 
-import com.exasol.mavenpluginintegrationtesting.MavenIntegrationTestEnvironment;
 import com.exasol.projectkeeper.config.ProjectKeeperConfig;
 import com.exasol.projectkeeper.config.ProjectKeeperConfigWriter;
-import com.exasol.projectkeeper.validators.TestMavenModel;
 
 public class ProjectKeeperAbstractIT {
     @TempDir
     protected Path projectDir;
-    private static MavenIntegrationTestEnvironment mavenIntegrationTestEnvironment;
-
-    @BeforeAll
-    static void beforeAll() {
-        mavenIntegrationTestEnvironment = TestEnvBuilder.getTestEnv();
-    }
 
     @BeforeEach
     void beforeEach() throws GitAPIException {
         Git.init().setDirectory(this.projectDir.toFile()).call().close();
     }
 
-    protected void writePomWithOneDependency(final String pomVersion, final String dependencyVersion)
-            throws IOException {
-        final TestMavenModel testMavenModel = new TestMavenModel();
-        testMavenModel.setVersion(pomVersion);
-        testMavenModel.addDependency("error-reporting-java", "com.exasol", "compile", dependencyVersion);
-        testMavenModel.configureAssemblyPluginFinalName();
-        testMavenModel.writeAsPomToProject(this.projectDir);
-    }
-
-    public Path getMavenRepo() {
-        return mavenIntegrationTestEnvironment.getLocalMavenRepository();
+    protected Optional<Path> getMavenRepo() {
+        return Optional.empty();
     }
 
     protected void runFix() {
@@ -59,15 +41,15 @@ public class ProjectKeeperAbstractIT {
         return logger.toString();
     }
 
-    public void writeConfig(final ProjectKeeperConfig config) {
+    protected void writeConfig(final ProjectKeeperConfig config) {
         new ProjectKeeperConfigWriter().writeConfig(config, this.projectDir);
     }
 
     protected ProjectKeeper getProjectKeeper(final Logger logger) {
-        return ProjectKeeper.createProjectKeeper(logger, this.projectDir, getMavenRepo(), CURRENT_VERSION);
+        return ProjectKeeper.createProjectKeeper(logger, this.projectDir, getMavenRepo().orElse(null), CURRENT_VERSION);
     }
 
-    public void assertVerifySucceeds() {
+    protected void assertVerifySucceeds() {
         final boolean success = getProjectKeeper(mock(Logger.class)).verify();
         assertThat(success, equalTo(true));
     }
