+++ project-keeper/src/test/java/com/exasol/projectkeeper/ProjectKeeperGolangIT.java
@@ -1,0 +1,8 @@
+package com.exasol.projectkeeper;
+
+import org.junit.jupiter.api.Tag;
+
+@Tag("integration")
+public class ProjectKeeperGolangIT {
+
+}
